Add utils files, remove lint errors, general cleanup
