+++ app/packages/multimodal/src/adapters/mcap/react/use-interpolated-image-annotations.ts
@@ -8,6 +8,7 @@ import type {
   ImageAnnotationText,
   ImageAnnotationsVisualization,
 } from "../../../decoders";
+import { groupLineSegmentsByLabel } from "../../../visualization/panels/line-list-grouping";
 import type { McapDecodedMessage } from "../types";
 import { useMcapDataStream } from "./mcap-data-stream-context";
 
@@ -276,38 +277,13 @@ interface Group {
   readonly label: string | null;
 }
 
-/**
- * Each annotation message encodes N objects with cuboid edges and labels
- * paired by index: `points` is `N * segmentsPerObject * 2` long and
- * `texts` has length N. The Nth chunk of segments is labeled by the Nth
- * text. We use that as the source of truth instead of spatial guessing.
- * Falls back to one big group when the data doesn't divide cleanly.
- */
 function groupLineList(
   points: readonly Point2[],
   texts: readonly ImageAnnotationText[]
 ): readonly Group[] {
-  const segmentCount = Math.floor(points.length / 2);
-  if (segmentCount === 0) return [];
-  if (texts.length === 0 || segmentCount % texts.length !== 0) {
-    const segments: [Point2, Point2][] = [];
-    for (let i = 0; i < segmentCount; i++) {
-      segments.push([points[i * 2], points[i * 2 + 1]]);
-    }
-    return [makeGroup(segments, null)];
-  }
-  const segmentsPerObject = segmentCount / texts.length;
-  const groups: Group[] = [];
-  for (let i = 0; i < texts.length; i++) {
-    const segments: [Point2, Point2][] = [];
-    const start = i * segmentsPerObject;
-    for (let j = 0; j < segmentsPerObject; j++) {
-      const seg = start + j;
-      segments.push([points[seg * 2], points[seg * 2 + 1]]);
-    }
-    groups.push(makeGroup(segments, texts[i]?.text ?? null));
-  }
-  return groups;
+  return groupLineSegmentsByLabel(points, texts).map(({ label, segments }) =>
+    makeGroup(segments, label)
+  );
 }
 
 function makeGroup(
