+++ app/packages/multimodal/src/load-status.ts
@@ -1,0 +1,7 @@
+/**
+ * Shared load-state union used by React hooks throughout the multimodal
+ * package. Each hook re-exports a domain-specific alias (e.g.
+ * `McapTopicsStatus`, `TemporalTagsStatus`) so consumers keep stable,
+ * readable names while the underlying definition stays in one place.
+ */
+export type LoadStatus = "idle" | "loading" | "ready" | "error";
